Added RoadmapDAO test, improved methods, fixed queries

+++ src/dao/RoadmapDAO.java
@@ -28,7 +28,7 @@ public class RoadmapDAO extends DataAccesObject{
 			ResultSet result = statement.executeQuery();
 			
 			while (result.next()) {
-				Roadmap roadmap = new Roadmap(result.getInt("roadmap_id"), result.getString("roadmapName"), result.getString("roadmapDescription"));
+				Roadmap roadmap = new Roadmap(result.getInt("roadmap_id"), result.getString("name"), result.getString("description"));
 				if (!theRoadmaps.contains(roadmap)) {
 					CategoryDAO categoryDAO = new CategoryDAO();
 					for(Category category : categoryDAO.getCategoriesByRoadmap(roadmap)) {
@@ -61,11 +61,10 @@ public class RoadmapDAO extends DataAccesObject{
 	public List<Roadmap> getAllRoadmapsByMentor(Mentor mentor) {
 		List<Roadmap> theRoadmaps = new ArrayList<Roadmap>();
 		try {
-			statement = con.prepareStatement("SELECT Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Step_has_Child.completed ,Achievement.achievement_id, Achievement.name as achievementName, Achievement.points"
-					+ "FROM Roadmap"
-					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id"
-					+ "JOIN Step_has_Child ON Step_has_Child.step_id = Step.step_id"
-					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id"
+			statement = con.prepareStatement("SELECT Roadmap.roadmap_id, Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Achievement.achievement_id, Achievement.name as achievementName, Achievement.points "
+					+ "FROM Roadmap "
+					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id "
+					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id "
 					+ "WHERE Roadmap.mentor_id = ?;");
 			statement.setInt(1, mentor.getMentorId());
 						
@@ -104,12 +103,11 @@ public class RoadmapDAO extends DataAccesObject{
 	public List<Roadmap> getAllRoadmapsByChild(Child child) {
 		List<Roadmap> theRoadmaps = new ArrayList<Roadmap>();
 		try {
-			statement = con.prepareStatement("SELECT Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Step_has_Child.completed ,Achievement.achievement_id, Achievement.name as achievementName, Achievement.points"
-					+ "FROM Roadmap"
-					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id"
-					+ "JOIN Step_has_Child ON Step_has_Child.step_id = Step.step_id"
-					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id"
-					+ "JOIN Child_has_Roadmap ON Roadmap.roadmap_id = Child_has_Roadmap.roadmap_id"
+			statement = con.prepareStatement("SELECT Roadmap.roadmap_id, Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Achievement.achievement_id, Achievement.name as achievementName, Achievement.points "
+					+ "FROM Roadmap "
+					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id "
+					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id "
+					+ "JOIN Child_has_Roadmap ON Roadmap.roadmap_id = Child_has_Roadmap.roadmap_id "
 					+ "WHERE Child_has_Roadmap.child_id = ?");
 			statement.setInt(1, child.getChildId());
 						
@@ -148,12 +146,11 @@ public class RoadmapDAO extends DataAccesObject{
 	public List<Roadmap> getAllRoadmapsByCategory(Category category) {
 		List<Roadmap> theRoadmaps = new ArrayList<Roadmap>();
 		try {
-			statement = con.prepareStatement("SELECT Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Step_has_Child.completed ,Achievement.achievement_id, Achievement.name as achievementName, Achievement.points"
-					+ "FROM Roadmap"
-					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id"
-					+ "JOIN Step_has_Child ON Step_has_Child.step_id = Step.step_id"
-					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id"
-					+ "JOIN Category_has_Roadmap ON Category_has_Roadmap.roadmap_id = Roadmap.roadmap_id"
+			statement = con.prepareStatement("SELECT Roadmap.roadmap_id, Roadmap.name AS roadmapName, Roadmap.description AS roadmapDescription, Step.step_id, Step.order_id as orderID, Step.name as stepName, Step.description as stepDescription, Achievement.achievement_id, Achievement.name as achievementName, Achievement.points " 
+					+ "FROM Roadmap " 
+					+ "JOIN Step ON Roadmap.roadmap_id = Step.roadmap_id "
+					+ "JOIN Achievement ON Roadmap.achievement_id = Achievement.achievement_id " 
+					+ "JOIN Category_has_Roadmap ON Category_has_Roadmap.roadmap_id = Roadmap.roadmap_id "
 					+ "WHERE Category_has_Roadmap.category_id = ?");
 			statement.setInt(1, category.getCategoryId());
 						
@@ -163,12 +160,12 @@ public class RoadmapDAO extends DataAccesObject{
 				if (!theRoadmaps.contains(roadmap)) {
 					Achievement achievement = new Achievement(result.getInt("achievement_id"), result.getString("achievementName"), result.getDouble("points"));
 					roadmap.setAchievement(achievement);
-					Step step = new Step(result.getInt("step_id"), result.getInt("orderID"), result.getString("stepName"), result.getString("stepDescription"), result.getBoolean("completed"));
+					Step step = new Step(result.getInt("step_id"), result.getInt("orderID"), result.getString("stepName"), result.getString("stepDescription"));
 					roadmap.addStep(step);
 					theRoadmaps.add(roadmap);
 				} else {
 					for (Roadmap r : theRoadmaps){
-						Step step = new Step(result.getInt("step_id"), result.getInt("orderID"), result.getString("stepName"), result.getString("stepDescription"), result.getBoolean("completed"));
+						Step step = new Step(result.getInt("step_id"), result.getInt("orderID"), result.getString("stepName"), result.getString("stepDescription"));
 						r.getSteps().add(step);
 					}
 				}
@@ -189,13 +186,13 @@ public class RoadmapDAO extends DataAccesObject{
 	public boolean addRoadmap(Roadmap roadmap) {
 		boolean succes = false;
 		try {
-			statement = con.prepareStatement("INSERT INTO Roadmap (`roadmap_id`, `name`, `description`, `mentor_id`, `achievement_id`) VALUES (NULL, ?, ?, ?, ?);");
+			statement = con.prepareStatement("INSERT INTO Roadmap (`name`, `description`, `mentor_id`, `achievement_id`) VALUES (?, ?, ?, ?);");
 			statement.setString(1, roadmap.getName());
 			statement.setString(2, roadmap.getDescription());
 			statement.setInt(3, roadmap.getMentor().getMentorId());
 			statement.setInt(4, roadmap.getAchievement().getId());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				for(Step step : roadmap.getSteps()) {
 					addStep(step, roadmap);
 				}
@@ -225,7 +222,7 @@ public class RoadmapDAO extends DataAccesObject{
 			statement.setInt(3, roadmap.getMentor().getMentorId());
 			statement.setInt(4, roadmap.getAchievement().getId());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				for(Step step : roadmap.getSteps()) {
 					updateStep(step);
 				}
@@ -248,15 +245,16 @@ public class RoadmapDAO extends DataAccesObject{
 	public boolean deleteRoadmap(Roadmap roadmap) {
 		boolean succes = false;
 		try {
-			statement = con.prepareStatement("DELETE FROM Roadmap WHERE Roadmap.roadmap_id = ?");
-			statement.setInt(1, roadmap.getId());
-			
-			for(Step step : roadmap.getSteps()) {
-				deleteStep(step);
-			}
-			
-			if(statement.execute() == true) {
-				succes = deleteChildHasRoadmap(roadmap) && deleteCategoryHasRoadmap(roadmap) && deleteRoadmapHasChild(roadmap);
+			if(deleteChildHasRoadmap(roadmap) && deleteCategoryHasRoadmap(roadmap)) {
+				for(Step step : roadmap.getSteps()) {
+					deleteStep(step);
+				}
+				
+				statement = con.prepareStatement("DELETE FROM Roadmap WHERE Roadmap.roadmap_id = ?");
+				statement.setInt(1, roadmap.getId());
+				if(statement.executeUpdate() > 0) {
+					succes = true;
+				}
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -279,10 +277,10 @@ public class RoadmapDAO extends DataAccesObject{
 			statement.setInt(1, child.getChildId());
 			statement.setInt(2, roadmap.getId());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				for(Step step : roadmap.getSteps()) {
 					addStepHasChild(step, child);
-				};
+				}
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -299,18 +297,19 @@ public class RoadmapDAO extends DataAccesObject{
 		return succes;
 	}
 	
-	public boolean deleteRoadmapHasChild(Roadmap roadmap) {
+	public boolean addRoadmapHasCategory(Roadmap roadmap, Category category) {
 		boolean succes = false;
 		try {
-			statement = con.prepareStatement("DELETE FROM `storytime`.`Child_has_Roadmap` WHERE `Child_has_Roadmap`.`roadmap_id` = ?");
-			statement.setInt(1, roadmap.getId());
+			statement = con.prepareStatement("INSERT INTO `storytime`.`Category_has_Roadmap` (`category_id`, `roadmap_id`) VALUES (?, ?);");
+			statement.setInt(1, category.getCategoryId());
+			statement.setInt(2, roadmap.getId());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
-			log.out(Level.ERROR, "", "Deleting Child_has_Roadmap went wrong");
+			log.out(Level.ERROR, "", "Adding Category_has_Roadmap went wrong");
 		} finally {
 			try {
 				statement.close();
@@ -332,7 +331,7 @@ public class RoadmapDAO extends DataAccesObject{
 		try {
 			statement = con.prepareStatement("DELETE FROM `storytime`.`Child_has_Roadmap` WHERE `Child_has_Roadmap`.`roadmap_id` = ?");
 			statement.setInt(1, roadmap.getId());
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -359,7 +358,7 @@ public class RoadmapDAO extends DataAccesObject{
 		try {
 			statement = con.prepareStatement("DELETE FROM `storytime`.`Category_has_Roadmap` WHERE `Category_has_Roadmap`.`roadmap_id` = ?");
 			statement.setInt(1, roadmap.getId());
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -417,7 +416,7 @@ public class RoadmapDAO extends DataAccesObject{
 			statement.setString(3, step.getDescription());
 			statement.setInt(4, roadmap.getId());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -438,12 +437,12 @@ public class RoadmapDAO extends DataAccesObject{
 		boolean succes = false;
 		try {
 			statement = con.prepareStatement("UPDATE `storytime`.`Step` SET `order_id` = ?, `name` = ?, `description` = ? WHERE `Step`.`step_id` = ?;");
-			statement.setInt(5, step.getId());
+			statement.setInt(4, step.getId());
 			statement.setInt(1, step.getOrderID());
 			statement.setString(2, step.getName());
 			statement.setString(3, step.getDescription());
 			
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -468,10 +467,14 @@ public class RoadmapDAO extends DataAccesObject{
 	public boolean deleteStep(Step step) {
 		boolean succes = false;
 		try {
-			statement = con.prepareStatement("DELETE FROM `storytime`.`Roadmap` WHERE `Roadmap`.`roadmap_id` = ?");
-			statement.setInt(1, step.getId());
-			if(statement.execute() == true) {
-				succes = deleteStepHasChild(step) && deleteStepHasChild(step);
+			if(deleteStepHasChild(step)) {
+				statement = con.prepareStatement("DELETE FROM `storytime`.`Step` WHERE `Step`.`step_id` = ?");
+				statement.setInt(1, step.getId());
+				if(statement.executeUpdate() > 0) {
+					succes = true;
+				}
+			} else {
+				succes = false;
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
@@ -493,7 +496,7 @@ public class RoadmapDAO extends DataAccesObject{
 			statement = con.prepareStatement("INSERT INTO `storytime`.`Step_has_Child` (`step_id`, `child_id`, `completed`) VALUES (?, ?, NULL);");
 			statement.setInt(1, step.getId());
 			statement.setInt(2, child.getChildId());
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -520,7 +523,7 @@ public class RoadmapDAO extends DataAccesObject{
 		try {
 			statement = con.prepareStatement("DELETE FROM `storytime`.`Step_has_Child` WHERE `Step_has_Child`.`step_id` = ?");
 			statement.setInt(1, step.getId());
-			if(statement.execute() == true) {
+			if(statement.executeUpdate() > 0) {
 				succes = true;
 			}
 		} catch (SQLException e) {
@@ -535,5 +538,25 @@ public class RoadmapDAO extends DataAccesObject{
 			}
 		}
 		return succes;
+	}
+	
+	// For testing purpose
+	public int getLatestIdRoadmap() {
+		int questionId = 0;
+		try {
+			statement = con.prepareStatement("SELECT MAX(roadmap_id) FROM Roadmap");
+			ResultSet result = statement.executeQuery();
+			result.next();
+			questionId = result.getInt(1);
+		} catch (SQLException e) {
+			e.printStackTrace();
+		} finally {
+			try {
+				statement.close();
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
+		}
+		return questionId;
 	}
 }
